Chef and Happy String
Problem:

Chef has a string S with him. Chef is happy if the string contains a contiguous substring of length strictly greater than 2 in which all its characters are vowels.

Determine whether Chef is happy or not.

Note that, in English alphabet, vowels are a, e, i, o, and u.