Fitness.java
Problem:

Chef wants to become fit for which he decided to walk to the office and return home by walking. It is known that Chef's office is X km away from his home.

If his office is open 5 days in a week, find the number of kilometers the Chef travels through office trips in a week.