Minimum Attendance Requirement.java
Problem:

Read problem statements in Hindi, Mandarin Chinese, Vietnamese, and Bengali as well.
A semester at Chef's University has 120 working days. The University requires that a student should be present for at least 75% of the working days in the semester. If not, the student is failed.

Chef has been taking a lot of holidays and is now concerned about whether he can pass the attendance requirement or not. N working days have already passed, and you are given N bits - B1​, B2​, ..., BN​.  Bi​ = 0 denotes that Chef was absent on the ℎith day, and Bi​ = 1 denotes that Chef was present on that day.

Can Chef hope to pass the requirement by the end of the semester?