Hardest Problem Bet.java
Problem:

There are 3 problems in a contest namely A,B,C respectively. Alice bets Bob that problem C is the hardest while Bob says that problem B will be the hardest.

You are given three integers SA, SB, and SC​ which denote the number of successful submissions of problems A,B,C respectively. It is guaranteed that each problem has a different number of submissions. Determine who wins the bet.

If Alice wins the bet (i.e. problem C is the hardest), then output Alice.
If Bob wins the bet (i.e. problem B is the hardest), then output Bob.
If no one wins the bet (i.e. problem A is the hardest), then output Draw.
Note: The hardest problem is the problem with the least number of successful submissions.