Good Quality Bulbs.java
Problem: 

A bulb company claims that the average lifetime of its bulbs is at least X units.

The company ran a test on N bulbs. You are given the lifetime of N−1 of these bulbs. What is the minimum non-negative integer value of the lifetime the remaining bulb can have, such that the claim of the company holds true?

problem link: https://www.codechef.com/problems/BULBLIFE