Passing Marks.java
Problem:

In a class of N students, a class test was held. The ith student scored Ai​ marks. It is also known that the scores of all students were distinct.

A student passes the test if their score is strictly greater than the passing mark. Given that exactly X students pass in the test, find the maximum value of the passing mark of the test.

problem link: https://www.codechef.com/problems/CUTOFF