Chef and Dolls.java
Problem:
 

Chef is a fan of pairs and he likes all things that come in pairs. He even has a doll collection in which the dolls come in pairs. One day while going through his collection he found that there are an odd number of dolls. Someone had stolen a doll!!!

Help the chef find which type of doll is missing.