How much Scholarship.java
Problem:

The ZCO Scholarship Contest has just finished, and you finish with a rank of R. You know that Rank 1 to Rank 50 will get a 100% scholarship on the ZCO exam fee and Rank 51 to Rank 100 will get a 50% percentage scholarship on the ZCO exam fee. The rest do not get any scholarships.
What percentage of the scholarship will you get?