Chef in Vaccination Queue.java
Problem:

There are N people in the vaccination queue, Chef is standing in the Pth position from the front of the queue. It takes X minutes to vaccinate a child and Y minutes to vaccinate an elderly person. Assume Chef is an elderly person.

You are given a binary array A1, A2,…, AN​ of length N, where Ai​=1 denotes there is an elderly person standing on the ith position of the queue, Ai​=0 denotes there is a child standing on the ith position of the queue. You are also given the three integers P, X, Y. Find the number of minutes after which the Chef's vaccination will be completed.