Airline Restrictions
Problem:

Read problem statements in Bengali, Mandarin Chinese, Russian, and Vietnamese as well.
Chef has 3 bags that she wants to take on a flight. They weigh A, B, and C kgs respectively. She wants to check in exactly two of these bags and carry the remaining one bag with her.

The airline restrictions say that the total sum of the weights of the bags that are checked-in cannot exceed D kgs and the weight of the bag which is carried cannot exceed E kgs. Find out if the Chef can take all three bags on the flight.

problem link: https://www.codechef.com/problems/AIRLINE