From heaven to earth.py
Problem:

The chef has been working in a restaurant which has N floors. He wants to minimize the time it takes him to go from the N-th floor to the ground floor. He can either take the elevator or the stairs.

The stairs are at an angle of 45 degrees and Chef's velocity is V1​ m/s when taking the stairs down. The elevator on the other hand moves with a velocity V2​ m/s. Whenever an elevator is called, it starts from the ground floor and goes to the N-th floor where it collects Chef (collecting takes no time), and makes its way down to the ground floor with Chef in it.

The elevator covers a total distance equal to N meters when going from the N-th floor to the ground floor or vice versa, while the length of the stairs is N∗2​ because the stairs are at an angle of 45 degrees.

Chef has enlisted your help to decide whether he should use the stairs or the elevator to minimize his travel time. Can you help him out?

Porblem  link: https://www.codechef.com/problems/ELEVSTRS?tab=statement