Binary String Cost.py
Problem:

According to the new tax scheme, a new tax called the binary string tax was introduced! JJ has a binary string S of length N. According to the tax scheme, for every occurrence of 01 in S, a tax of X rupees will be charged, while for every occurrence of 10 in S, a tax of Y rupees will be charged.

For example,X=5,  Y=7 and S=11010, then S has 1 occurrence of 01 and 2 occurrences of 10, so the tax charged 5⋅1+7⋅2=19

JJ can rearrange the string S in any way he wants. He wants to minimize the amount of tax he has to pay. Can you help him do so?

Porblem  link: https://www.codechef.com/problems/BSCOST