Card Swipe
Problem:

In the bustling CodeChef office, the entrance is equipped with a high-tech card swipe system. Each employee is assigned a unique ID card that they use to swipe in and out of the building.

The system records every swipe, capturing the first swipe of an ID as in, second as out, third as in, and so on. Given an array A consisting of N IDs denoting N swipes throughout the day, find the maximum number of people in the office at any time.

Note that there is nobody inside the office before the first swipe.

Problem link:  https://www.codechef.com/practice/course/1-star-difficulty-problems/DIFF1200/problems/CARDSWIPE?tab=statement