Chef and Groups
Problem:

There are N seats in a row. You are given a string S with length N; for each valid i, the i-th character of S is '0' if the i-th seat is empty or '1' if there is someone sitting in that seat.

Two people are friends if they are sitting next to each other. Two friends are always part of the same group of friends. Can you find the total number of groups?

Problem link:  https://www.codechef.com/practice/course/1-star-difficulty-problems/DIFF1200/problems/GROUPS?tab=statement