Football Ties
Football Ties

According to the Premier League rules, the winner of a game gets 3 points while the loser of a game gets 0 points. On the other hand, if there is a draw, both sides get 1 point each.

Two teams A and B played several games against each other and finally ended up with the score of X and Y points respectively.
We do not know how many games were played. Find the minimum possible number of draws that may have happened.

It is guaranteed that there is at least one way to end up with the final score being X and Y points respectively. Football Ties
According to the Premier League rules, the winner of a game gets 3 points while the loser of a game gets 0 points. On the other hand, if there is a draw, both sides get 1 point each.

Two teams A and B played several games against each other and finally ended up with the score of X and Y points respectively.
We do not know how many games were played. Find the minimum possible number of draws that may have happened.

It is guaranteed that there is at least one way to end up with the final score being X and Y points respectively.

Problem link: https://www.codechef.com/START135D/problems/FOOTBALLTIES