Ping Paja Pong
Problem: 

Chef and Paja are bored, so they are playing an infinite game of ping pong. The rules of the game are as follows:

The players play an infinite number of games. At the end of each game, the player who won it scores a point.
In each game, one of the players serves. Chef serves in the first game.
After every K points are scored (regardless of which players scored them), i.e. whenever K games have been played since the last time the serving player changed, the player that serves in the subsequent games changes: if Chef served in the game that just finished, then Paja will serve in the next game and all subsequent games until the serving player changes again; if Paja served, then Chef will serve.
The players got a little too caught up in the game and they forgot who is supposed to serve in the next game. Will you help them determine that? So far, Chef has scored X points and Paja has scored Y points.

Problem link: 
https://www.codechef.com/problems/PAJAPONG