Palindrome Flipping
Palindrome Flipping

Problem: 

Chef has a binary string 𝑆S of length 𝑁N.

In one operation, Chef can:

Select two indices i and j (1≤i,j≤N,i!=j) and flip Si​ and Sj​. (i.e. change 0 to 1 and 1 to 0)
For example, if S=10010 and chef applys operation on i=1 and j=3 then: 1​00​10→00110.

Find if it is possible to convert 𝑆S to a palindrome by applying the above operation any (possibly zero) number of times.

Note: A string is called a palindrome if it reads the same backward and forwards, e.g. 10001 and 0110 are palindromic strings.

Problem link: 
https://www.codechef.com/problems/FLIPPAL