Non-Negative Product
Problem: 

Alice has an array of N integers —𝐴𝑁A1​,A2​,…,AN​. She wants the product of all the elements of the array to be a non-negative integer. That is, it can be either 0 or positive. But she doesn't want it to be negative.

To do this, she is willing to remove some elements of the array. Determine the minimum number of elements that she will have to remove to make the product of the array's elements non-negative.

Problem link: 
https://www.codechef.com/problems/NONNEGPROD