Problem:
The Hamming distance between two pairs of strings of equal length is defined to be the number of positions at which they contain different characters.
For example, the Hamming distance between strings "there" and "shire" is 2 (their first and third characters are different), while the Hamming distance between "order" and "chaos" is 5, since they differ at every position.

Chef has a string A of length N.
Chef's favorite string is B, which has length M. It is known that M≤N.

Find the minimum Hamming distance between B and some contiguous substring of A that has length M.

A substring of a string is obtained by deleting some (possibly, zero) characters from its beginning and some (possibly, zero) characters from its end.
For example, "abc", "bc", and "cd" are substrings of "abcd", but "ac" is not.

Problem link:
https://www.codechef.com/problems/SAMESAME